[CHORE] Cleanup
